Updated a lot of dependencies.

+++ src/dns/protocol.rs
@@ -973,7 +973,7 @@ impl DnsPacket {
 
     pub fn get_random_a(&self) -> Option<String> {
         if !self.answers.is_empty() {
-            let idx = random::<usize>() % self.answers.len();
+            let idx = random::<u64>() as usize % self.answers.len();
             let a_record = &self.answers[idx];
             if let DnsRecord::A { ref addr, .. } = *a_record {
                 return Some(addr.to_string());
@@ -1044,7 +1044,7 @@ impl DnsPacket {
         }
 
         if !new_authorities.is_empty() {
-            let idx = random::<usize>() % new_authorities.len();
+            let idx = random::<u64>() as usize % new_authorities.len();
             if let DnsRecord::A { addr, .. } = new_authorities[idx] {
                 return Some(addr.to_string());
             }
@@ -1066,7 +1066,7 @@ impl DnsPacket {
         }
 
         if !new_authorities.is_empty() {
-            let idx = random::<usize>() % new_authorities.len();
+            let idx = random::<u64>() as usize % new_authorities.len();
             return Some(new_authorities[idx].clone());
         }
